Update the scripts

+++ Scripts/clean-branch.ps1
@@ -1,7 +1,7 @@
 #!/bin/powershell
 <#
 .SYNTAX         ./clean-branch.ps1 [<repo-dir>]
-.DESCRIPTION	cleans the current Git branch including submodules from generated files (e.g. for a fresh build)
+.DESCRIPTION	cleans the current/given Git repository from generated files (including submodules, e.g. for a fresh build)
 .LINK		https://github.com/fleschutz/PowerShell
 .NOTES		Author:	Markus Fleschutz / License: CC0
 #>
@@ -16,8 +16,10 @@ try {
 }
 
 try {
-	write-progress "Cleaning current branch in repository $RepoDir..."
-	set-location $RepoDir
+	write-progress "Cleaning Git repository $RepoDir from generated files ..."
+
+	if (-not(test-path "$RepoDir" -pathType container)) { throw "Can't access directory: $RepoDir" }
+	set-location "$RepoDir"
 	
 	& git clean -fdx # force + recurse into dirs + don't use the standard ignore rules
 	if ($lastExitCode -ne "0") { throw "'git clean -fdx' failed" }
@@ -25,7 +27,7 @@ try {
 	& git submodule foreach --recursive git clean -fdx 
 	if ($lastExitCode -ne "0") { throw "'git clean -fdx' in submodules failed" }
 
-	write-host -foregroundColor green "OK - cleaned current branch in repository $RepoDir"
+	write-host -foregroundColor green "OK - cleaned Git repository $RepoDir from generated files"
 	exit 0
 } catch {
 	write-error "ERROR: line $($_.InvocationInfo.ScriptLineNumber): $($Error[0])"
